requirements.txt
fastapi
uvicorn
aiogram
supabase
jinja2
python-dotenv
python-multipart
google-genai